Bayesian Emax model fit with Stan

Run sigmoidal Emax model fit with formula notation

stan_emax(formula, data, gamma.fix = 1, e0.fix = NULL, priors = NULL,
  ...)

Arguments

formula
a symbolic description of variables for Emax model fit.

data
an optional data frame containing the variables in the model.

gamma.fix
a numeric or NULL to specify gamma (Hill coefficient) in the sigmoidal Emax model. If NULL, gamma will be estimated from the data. If numeric, gamma is fixed at the number provided. Default = 1 (normal Emax model).

e0.fix
a numeric or NULL to specify E0 in the Emax model. If NULL, E0 will be estimated from the data. If numeric, E0 is fixed at the number provided. Default = NULL (estimate from the data).

priors
a named list specifying priors of parameters (ec50, emax, e0, gamma, sigma). Each list item should be length 2 numeric vector, one corresponding to mean and another corresponding to standard deviation. Currently only supports normal distribution for priors.

...
Arguments passed to `rstan::sampling` (e.g. iter, chains).

Value

An object of class `stanemax`

Examples

data(exposure.response.sample)
fit1 <- stan_emax(response ~ exposure, data = exposure.response.sample,
                  # the next line is only to make the example go fast enough
                  chains = 1, iter = 500, seed = 12345)

#> 
#> SAMPLING FOR MODEL 'emax' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.002 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 20 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 500 [  0%]  (Warmup)
#> Chain 1: Iteration:  50 / 500 [ 10%]  (Warmup)
#> Chain 1: Iteration: 100 / 500 [ 20%]  (Warmup)
#> Chain 1: Iteration: 150 / 500 [ 30%]  (Warmup)
#> Chain 1: Iteration: 200 / 500 [ 40%]  (Warmup)
#> Chain 1: Iteration: 250 / 500 [ 50%]  (Warmup)
#> Chain 1: Iteration: 251 / 500 [ 50%]  (Sampling)
#> Chain 1: Iteration: 300 / 500 [ 60%]  (Sampling)
#> Chain 1: Iteration: 350 / 500 [ 70%]  (Sampling)
#> Chain 1: Iteration: 400 / 500 [ 80%]  (Sampling)
#> Chain 1: Iteration: 450 / 500 [ 90%]  (Sampling)
#> Chain 1: Iteration: 500 / 500 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.479 seconds (Warm-up)
#> Chain 1:                0.176 seconds (Sampling)
#> Chain 1:                0.655 seconds (Total)
#> Chain 1: 

print(fit1)

#> Inference for Stan model: emax.
#> 1 chains, each with iter=500; warmup=250; thin=1; 
#> post-warmup draws per chain=250, total post-warmup draws=250.
#> 
#>         mean se_mean    sd  2.5%    25%    50%    75%  97.5% n_eff Rhat
#> emax  107.33    0.32  4.65 98.67 103.91 106.99 110.80 116.55   212    1
#> e0      6.23    0.27  3.71 -0.33   3.66   6.15   8.87  14.12   190    1
#> ec50   98.72    1.56 20.47 67.31  83.27  95.45 111.43 144.57   172    1
#> gamma   1.00     NaN  0.00  1.00   1.00   1.00   1.00   1.00   NaN  NaN
#> sigma  19.99    0.09  1.32 17.64  19.12  19.91  20.78  23.15   204    1
#> 
#> Samples were drawn using NUTS(diag_e) at Tue May 28 22:02:22 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# Set priors manually, also estimate gamma instead of the default of fix to 1
fit2 <- stan_emax(response ~ exposure, data = exposure.response.sample, gamma.fix = NULL,
                  priors = list(ec50  = c(100, 30), emax  = c(100, 30), e0 = c(10, 5),
                                gamma = c(0, 3), sigma = c(0, 30)),
                  # the next line is only to make the example go fast enough
                  chains = 1, iter = 500, seed = 12345)

#> 
#> SAMPLING FOR MODEL 'emax' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 500 [  0%]  (Warmup)
#> Chain 1: Iteration:  50 / 500 [ 10%]  (Warmup)
#> Chain 1: Iteration: 100 / 500 [ 20%]  (Warmup)
#> Chain 1: Iteration: 150 / 500 [ 30%]  (Warmup)
#> Chain 1: Iteration: 200 / 500 [ 40%]  (Warmup)
#> Chain 1: Iteration: 250 / 500 [ 50%]  (Warmup)
#> Chain 1: Iteration: 251 / 500 [ 50%]  (Sampling)
#> Chain 1: Iteration: 300 / 500 [ 60%]  (Sampling)
#> Chain 1: Iteration: 350 / 500 [ 70%]  (Sampling)
#> Chain 1: Iteration: 400 / 500 [ 80%]  (Sampling)
#> Chain 1: Iteration: 450 / 500 [ 90%]  (Sampling)
#> Chain 1: Iteration: 500 / 500 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.44 seconds (Warm-up)
#> Chain 1:                0.135 seconds (Sampling)
#> Chain 1:                0.575 seconds (Total)
#> Chain 1: 

print(fit2)

#> Inference for Stan model: emax.
#> 1 chains, each with iter=500; warmup=250; thin=1; 
#> post-warmup draws per chain=250, total post-warmup draws=250.
#> 
#>         mean se_mean    sd  2.5%   25%    50%    75%  97.5% n_eff Rhat
#> emax  104.45    0.60  7.05 91.19 99.74 104.13 109.68 119.29   138 1.02
#> e0      8.05    0.30  3.14  1.48  5.89   8.14  10.51  13.32   113 1.00
#> ec50   99.25    1.33 16.78 71.46 86.49  97.63 109.85 134.01   160 1.00
#> gamma   1.07    0.02  0.17  0.79  0.95   1.05   1.17   1.45   122 1.02
#> sigma  20.08    0.09  1.35 17.65 19.16  20.05  20.87  22.65   207 1.00
#> 
#> Samples were drawn using NUTS(diag_e) at Tue May 28 22:02:25 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).